move tests to utils
